refactor(crouton-auth): break auth.ts ↔ user-profile.ts schema import cycle (#1240)
👤 For humans
The repo's one circular dependency is gone: the user table now lives in
its own schema file, so user-profile no longer has to import auth.ts
(which imports user-profile back). Nothing changes for consumers.

🤖 For agents
- New server/database/schema/user.ts: `user` table + User/NewUser types
- auth.ts imports + re-exports user (same public surface, all import
  paths keep working); user-profile.ts now imports from './user'
- Graph is now auth → {user, user-profile}, user-profile → user
- fallow dead-code: circular deps 1 → 0; package build green
