+++ packages/crouton-auth/server/database/schema/auth.ts
@@ -13,49 +13,20 @@
  */
 import { sqliteTable, text, integer, index } from 'drizzle-orm/sqlite-core'
 import { relations } from 'drizzle-orm'
+import { user } from './user'
 import { userProfile } from './user-profile'
 
-// Re-export user-profile schema so it's included in the built auth schema artifact
+// Re-export the user + user-profile schemas so they're included in the built
+// auth schema artifact (the user table lives in ./user so user-profile can
+// reference it without a circular import through this file)
+export { user } from './user'
+export type { User, NewUser } from './user'
 export { userProfile, userProfileRelations } from './user-profile'
 export type { UserProfileRecord, NewUserProfileRecord } from './user-profile'
 
 // ============================================================================
 // Core Tables
 // ============================================================================
-
-/**
- * User table - Core user identity
- *
- * Stores basic user information and profile data.
- * Extended with stripeCustomerId for billing support.
- */
-export const user = sqliteTable('user', {
-  id: text('id').primaryKey(),
-  name: text('name').notNull(),
-  email: text('email').notNull().unique(),
-  emailVerified: integer('emailVerified', { mode: 'boolean' }).notNull().default(false),
-  image: text('image'),
-  createdAt: integer('createdAt', { mode: 'timestamp' }).notNull().$default(() => new Date()),
-  updatedAt: integer('updatedAt', { mode: 'timestamp' }).notNull().$onUpdate(() => new Date()),
-  // Stripe extension
-  stripeCustomerId: text('stripeCustomerId'),
-  // Admin extension (crouton-admin package)
-  /** Whether this user has super admin privileges */
-  superAdmin: integer('superAdmin', { mode: 'boolean' }).notNull().default(false),
-  /** Whether this user is banned from the platform */
-  banned: integer('banned', { mode: 'boolean' }).notNull().default(false),
-  /** Reason for the ban (shown to user and admin) */
-  bannedReason: text('bannedReason'),
-  /** When the ban expires (null = permanent) */
-  bannedUntil: integer('bannedUntil', { mode: 'timestamp' }),
-  // Better Auth 1.5+ user role (admin plugin)
-  role: text('role').default('user')
-}, table => [
-  index('user_email_idx').on(table.email),
-  index('user_stripe_customer_idx').on(table.stripeCustomerId),
-  index('user_super_admin_idx').on(table.superAdmin),
-  index('user_banned_idx').on(table.banned)
-])
 
 /**
  * Session table - User sessions
@@ -548,9 +519,6 @@ export const authEmailLogRelations = relations(authEmailLog, ({ one }) => ({
 // ============================================================================
 // Type Exports
 // ============================================================================
-
-export type User = typeof user.$inferSelect
-export type NewUser = typeof user.$inferInsert
 
 export type Session = typeof session.$inferSelect
 export type NewSession = typeof session.$inferInsert
